myfgets OK

+++ directeur.h
@@ -5,4 +5,4 @@ pid_t creerProcessus();
 
 /*Appelle nombreDeProcessus processus fils qui appellent chef et
  * renvoient un resultat qui sera stocke dans le tableau resultats[] */
-void directeur(int resultats[], int nombreDeProcessus, char* argv[]);
+void directeur(double resultats[], int nombreDeProcessus, char* argv[]);
